refactor(query): extract query-model crate
