+++ crates/query-model/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "query-model"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+storage = { path = "../storage" }
+schema = { path = "../schema" }
+
+serde.workspace = true
+serde_json.workspace = true
+chrono = { version = "0.4", features = ["serde"] }
+tracing.workspace = true
+thiserror.workspace = true
