+++ crates/query-model/src/document/mod.rs
@@ -1,0 +1,3 @@
+mod mapping;
+
+pub use mapping::{DocumentMapping, RenderKey, DOC_ID_FIELD_INDEX};
