+++ crates/query-model/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod document;
+pub mod error;
+pub mod mapper;
