+++ crates/query/src/mapper/mod.rs
@@ -1,12 +1,3 @@
 //! Query mapper types for converting parsed queries to internal operations
 
-mod filter;
-mod mutation;
-mod types;
-
-pub use filter::{like_pattern_match, Filter, FilterOp};
-pub use mutation::{parse_mutation_name, Mutation, MutationType};
-pub use types::{
-    Aggregate, AggregateTarget, AggregateType, Field, FullTextSearch, GroupBy, Limit, OrderBy,
-    OrderCondition, OrderDirection, Requestable, Select, SelectionType, Similarity,
-};
+pub use query_model::mapper::*;
