+++ crates/query/src/runner/plan.rs
@@ -4,7 +4,7 @@ use std::sync::Arc;
 
 use acp::{DocumentACP, Identity};
 use schema::CollectionVersion;
-use serde_json::{Map, Value as JsonValue};
+use serde_json::Value as JsonValue;
 
 use crate::document::DocumentMapping;
 use crate::error::{QueryError, Result};
@@ -455,40 +455,5 @@ pub(crate) fn build_plan(
 /// Special handling for `__typename`: if the mapping has type_info set and the
 /// render_key index matches the __typename field index, the stored type name is used.
 pub(crate) fn doc_to_json(doc: &Doc, mapping: &DocumentMapping) -> Result<JsonValue> {
-    let mut obj = Map::new();
-
-    // Get the __typename index and name if set
-    let typename_info = mapping
-        .first_index_of_name("__typename")
-        .and_then(|idx| mapping.type_name().map(|name| (idx, name.to_string())));
-
-    // Get the _deleted index if present
-    let deleted_index = mapping.first_index_of_name("_deleted");
-
-    for render_key in &mapping.render_keys {
-        // Check for _deleted special handling
-        let value = if Some(render_key.index) == deleted_index && render_key.key == "_deleted" {
-            JsonValue::Bool(doc.is_deleted())
-        } else if let Some((typename_idx, ref typename)) = typename_info {
-            if render_key.index == typename_idx {
-                // Return the stored type name for __typename
-                JsonValue::String(typename.clone())
-            } else {
-                doc.fields()
-                    .get(render_key.index)
-                    .cloned()
-                    .flatten()
-                    .unwrap_or(JsonValue::Null)
-            }
-        } else {
-            doc.fields()
-                .get(render_key.index)
-                .cloned()
-                .flatten()
-                .unwrap_or(JsonValue::Null)
-        };
-        obj.insert(render_key.key.clone(), value);
-    }
-
-    Ok(JsonValue::Object(obj))
+    Ok(crate::document::render_doc_to_json(mapping, doc))
 }
